+++ changelog.md
@@ -8,6 +8,7 @@
 - Fix fluid router GUI crashing (Léa Gris)
 - Fix bullets not assigning their owner, thus all damage being considered turret damage (BluSunrize)
 - Fix bullets and railgun shots not counting as projectiles (BluSunrize)
+- Fix chemthrower not correctly attributing kills to the player (BluSunrize)
 
 ##### Version 1.21.1-12.1.0-185
 - All changes from 1.20.4-11.7.0